pyproject.toml
[tool.poetry]
name = "handbrainchess"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
python-chess = "^1.999"
Flask = "^2.0.2"
stockfish = "^3.18.0"
Flask-Cors = "^3.0.10"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
